realizado atualizações nos models de categorias e de usuários, e nos controllers de contas, categorias, transações e usuários

+++ src/app/controllers/account.js
@@ -10,7 +10,7 @@ router.use(auth)
 
 router.get('/', async (req, res) => {
   try {
-    const account = await Account.find().populate('usuario')
+    const account = await Account.find({ usuario: req.userId }).populate('usuario')
 
     return res.send(account)
   } catch (error) {
@@ -20,7 +20,7 @@ router.get('/', async (req, res) => {
 
 router.get('/:id', async (req, res) => {
   try {
-    const account = await Account.findById(req.params.id).populate('usuario')
+    const account = await Account.find({ id: req.params.id, usuario: req.userId }).populate('usuario')
 
     return res.send(account)
   } catch (error) {
@@ -47,6 +47,11 @@ router.post('/', async (req, res) => {
 router.put('/:id', async (req, res) => {
   try {
     const { descricao } = req.body
+    const { usuario } = await Account.findById(req.params.id)
+
+    if (usuario !== req.userId) {
+      return res.status(400).send({ error: 'Impossível atualizar conta bancária' })
+    }
 
     const account = await Account.findByIdAndUpdate(req.params.id, {
       descricao
@@ -64,8 +69,12 @@ router.put('/:id', async (req, res) => {
 
 router.delete('/:id', async (req, res) => {
   try {
-    const { saldo: saldoAccount } = await Account.findById(req.params.id)
+    const { saldo: saldoAccount, usuario } = await Account.findById(req.params.id)
     const { saldo: saldoUser } = await User.findById(req.userId)
+
+    if (usuario !== req.userId) {
+      return res.status(400).send({ error: 'Impossível atualizar conta bancária' })
+    }
 
     await Account.findByIdAndRemove(req.params.id)
